fix: corrected misleading test behaviour
fix: use UTC instead of local time

fix: corrected the inconsistentcy with validators

style: run prettier 😤 i swear i did it previously

fix: one more double parsing instead of uint

+++ backend/ResearchCruiseApp/Application/UseCases/CruiseApplications/AddCruiseApplication/AddCruiseApplicationHandler.cs
@@ -123,7 +123,7 @@ public class AddCruiseApplicationHandler(
         double cruiseDurationDays = 0;
         if (!string.IsNullOrWhiteSpace(formADto.CruiseHours))
         {
-            if (double.TryParse(formADto.CruiseHours, out var cruiseHours))
+            if (uint.TryParse(formADto.CruiseHours, out var cruiseHours))
             {
                 cruiseDurationDays = cruiseHours / 24.0;
             }
